perf(ci): optimize benchmark workflow from ~21min to ~5-8min
Three changes to reduce benchmark CI time:

1. Consolidate 5 bench binaries into 1: Each benchmark target required
   a full compilation+link cycle against the entire crate. Merging into
   a single binary with modules eliminates 4 redundant compilations
   (~8-12 min saved).

2. Add Criterion.toml with CI profile: Default criterion settings use
   3s warmup + 5s measurement per benchmark. The CI profile reduces
   this to 1s/2s with sample_size=20, still adequate for the 150%
   regression detection threshold (~5 min saved).

3. Merge format.rs and formatter.rs: These had duplicate benchmarks
   (both benchmarked format_table at same row counts). Consolidated
   into bench_format.rs keeping all unique benchmarks (~1 min saved).

Benchmark names in output.json are preserved (determined by
criterion_group! strings, not binary names), so the GitHub Pages
historical dashboard remains intact.
