+++ .github/workflows/bench.yml
@@ -53,7 +53,9 @@ jobs:
             ${{ runner.os }}-cargo-bench-
 
       - name: Run benchmarks
-        run: cargo bench 2>&1 | tee bench_output.txt
+        env:
+          CRITERION_PROFILE: ci
+        run: cargo bench --bench benchmarks 2>&1 | tee bench_output.txt
 
       - name: Parse benchmark results
         run: python3 scripts/format_bench_summary.py < bench_output.txt >> "$GITHUB_STEP_SUMMARY" || echo "No benchmark results" >> "$GITHUB_STEP_SUMMARY"
